feat: add env flie, login, logout func

+++ src/pages/api/post/delete.js
@@ -1,15 +1,24 @@
 import { connectDB } from "@/util/database";
 const ObjectId = require("mongodb").ObjectId;
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(req, res) {
   if (req.method === "POST") {
     const client = await connectDB;
     const db = client.db("next-board");
 
-    console.log(req);
+    let session = await getServerSession(req, res, authOptions);
 
-    await db.collection("board").deleteOne({ _id: new ObjectId(req.body) });
+    let data = await db
+      .collection("board")
+      .findOne({ _id: new ObjectId(req.body) });
 
-    res.status(200).json("success !");
+    if (data.author === session.user.name) {
+      await db.collection("board").deleteOne({ _id: new ObjectId(req.body) });
+      res.status(200).json("success !");
+    } else {
+      res.status(500).json("user error : can't delete this content !");
+    }
   }
 }
